Add options attribute to pass a complete options list to directive
The option attribute must contains an object with the additional options
you want to pass to the justgage object. Normal option are still use but
will be overload by object if declared twice.

Here's a sample to use levelColors and levelGradiant.
All justgage option are now available for the directive.

        <just-gage
            id="demo"
            class=""
            min=0 max=100 value="92"
            title="Demo"
            options="{
                levelColors: ['#CE1B21', '#D0532A', '#FFC414', '#85A137']
                ,levelGradiant: false
            }
            ">
        </just-gage>